chore: 유저 관련 API 분류
